requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==21.2.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
bleach==6.1.0
python-magic==0.4.27
requests==2.32.3
Pillow==11.3.0
openai>=1.99.0
